Add a reupload_if_updated option (default: true)
Following a discussion on feed2imap-devel, add a reupload_if_updated
option. When set to false, if an item is updated, but was previously
removed from the IMAP server, it is no longer re-uploaded.

Also fix some config file parsing bugs for the disable-ssl-verification
and include-images options. I should really switch to another way to
describe F2I config...